pyproject.toml
[tool.poetry]
name = "curatorsupportsystem"
version = "0.1.0"
description = ""
authors = ["Sofia <sonechka68zvr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.41"  # Или "^2.0.0", но не оба
pydantic = "^2.11.4"     # Или "^2.0.0", но не оба
fastapi = "^0.115.12"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
python-multipart = "0.0.19"
python-dotenv = "^1.0.0"
passlib = "^1.7.4"
schemas = "^0.7.1"
crud = "^0.1"
python-jose = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
